타 모듈 Bean 불러오기 오류 수정, Reservation 수정

+++ api-reservation/src/main/java/com/example/apireservation/reservation/ReservationService.java
@@ -2,6 +2,7 @@ package com.example.apireservation.reservation;
 
 import com.example.apireservation.reservation.model.Reservation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,8 +17,21 @@ public class ReservationService {
     public void save (Reservation reservation) throws Exception {
         reservationRepository.save(reservation);
     }
-    public void delete (Long idx) throws Exception {
-        reservationRepository.deleteById(idx);
+    public int delete (Long idx)  {
+        Optional<Reservation> reservation = reservationRepository.findById(idx);
+
+        if (!reservation.isPresent()) {
+            // 존재하지 않는 예약일 경우 예외를 던짐
+            return 1; // 예를 들어, 존재하지 않는 경우는 1 반환
+        }
+        try {
+            reservationRepository.deleteById(idx);
+            return 0;
+        } catch (EmptyResultDataAccessException e) {
+            return 1;
+        } catch (Exception e) {
+            return 2;
+        }
     }
     public Reservation findByIdx(Long idx) throws Exception {
         Optional<Reservation> reservationOptional = reservationRepository.findById(idx);
